backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pandas>=2.0.0
google-generativeai>=0.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
